+++ v0.2/GeneticAlgoritms/src/utils/Ponteiro.java
@@ -1,7 +1,3 @@
-/*
- * To change this template, choose Tools | Templates
- * and open the template in the editor.
- */
 package utils;
 
 import genetics.Individual;
@@ -10,22 +6,33 @@ import operators.Genetic;
 import utils.exceptions.PonteiroForaDoLimiteException;
 
 /**
- *
+ * Classe Ponteiro que será um ponteiro para a população. Este ponteiro será calculado aleatóriamente e apontará
+ * para um individuo da população passada por parametro.
  * @author PedroIsi
  */
 public class Ponteiro {
     
-    //define um ponto aleatorio entre zero e o total de fitness da população
-    public static double pontoAleatorio(double totalFitnessPopulacao){
-        return Genetic.RANDOM_GENERATOR.nextDouble() * totalFitnessPopulacao;
+    /**
+     * Método que devolve um ponto aleatorio entre 0 e o total de fitness da população
+     * @param totalFitnessPopulacao (int) - Total de fitness da população
+     * @return (double) - Valor apontado ao acaso pelo método (entre 0 e o fitness total da população)
+     */
+    public static double pontoAleatorio(int totalFitnessPopulacao){
+        return Genetic.RANDOM_GENERATOR.nextDouble() * (double)totalFitnessPopulacao;
     }
     
     //devolve individuo para onde o ponto aponta
+    /**
+     * Método que devolve o individuo para onde o ponteiro aponta
+     * @param ponteiro (double) - Valor entre 0 e o fitness total da população, sendo que este será o valor que irá apontar para um individuo em especifico
+     * @param population (Population) - População na qual será aplicado o ponteiro
+     * @return (Individual) - Individuo para o qual o ponteiro apontou
+     * @throws PonteiroForaDoLimiteException - Excepção caso o ponteiro aponte para "o vazio"
+     */
     public static Individual devolveIndividuoParaOndeOPonteiroAponta(double ponteiro, Population population) throws PonteiroForaDoLimiteException{
-        double totalFitness = 0;
-        
+        int totalFitness = 0;
         // verificar se o ponteiro é maior que o total de fitness da população ou
-        // se é menor que zero, se sim dispar uma excepção do tipo PonteiroForaDoLimiteException
+        // se é menor que zero; Se sim, dispara uma excepção do tipo PonteiroForaDoLimiteException
         if(ponteiro > PopulationUtils.getFitnessTotal(population) || ponteiro < 0) 
             throw new PonteiroForaDoLimiteException("Ponteiro dentro da class Ponteiro esta fora dos limites do total de fitness da população.");
         
@@ -35,13 +42,12 @@ public class Ponteiro {
             totalFitness += individuo.fitness();
             
             //escolhe o individuo onde o ponteiro aponta 
-            if(ponteiro <= totalFitness) {
+            if(ponteiro <= (double)totalFitness) {
                 //clone para criar um novo individuo e não ser individuo da pop original
                 return individuo.clone();
             }
         }
-        
+        //Caso não existam individuos na população recebida, devolve null
         return null;
     }
-    
 }
